Modularized css Styles
