+++ frontend/src/pages/Checkout/components/CarbonModal.module.css
@@ -1,0 +1,20 @@
+.container{
+    @apply fixed inset-0 flex items-center justify-center bg-black bg-opacity-50
+}
+
+.section{
+    @apply bg-white p-6 rounded-lg shadow-lg w-[900px] h-[500px] relative
+}
+
+.header{
+    @apply flex justify-between items-center
+}
+.heading{
+    @apply text-lg font-semibold text-center w-full
+}
+.closeButton{
+    @apply text-gray-500 hover:text-gray-700 text-2xl font-bold
+}
+.imgDiv{
+    @apply flex justify-center py-1
+}
