+++ frontend/src/pages/Checkout/components/EmissionMeter.jsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useState } from 'react';
 import Gauge from './Gauge';  
 import { useTransport } from '@/context/transport-context';
-import styles from "./OrderComponents.module.css";
+import styles from "./Emissionmeter.module.css";
 const EmissionMeter = ({ currentValue, maxValue, onEmissionsClick,setLowSustainable,  className = ""  }) => {
      
     // Normalize emissions for the gauge (200–1000 scale to 0–100 scale)
