+++ frontend/src/pages/Checkout/components/EmissionMeter.module.css
@@ -1,23 +1,3 @@
-.container{
-    @apply fixed inset-0 flex items-center justify-center bg-black bg-opacity-50
-}
-
-.section{
-    @apply bg-white p-6 rounded-lg shadow-lg w-[900px] h-[500px] relative
-}
-
-.header{
-    @apply flex justify-between items-center
-}
-.heading{
-    @apply text-lg font-semibold text-center w-full
-}
-.closeButton{
-    @apply text-gray-500 hover:text-gray-700 text-2xl font-bold
-}
-.imgDiv{
-    @apply flex justify-center py-1
-}
 
 .meterContainer{
     @apply flex items-center text-center justify-center p-4
@@ -38,4 +18,7 @@
 }
 .reasonLink{
     @apply mr-6 cursor-pointer
+}
+.imgDiv{
+    @apply flex justify-center py-1
 }
